Flushing metrics on ephemeral executions (#4291)
As things currently stand, metrics are flushed by a background thread
every 10m. However, for the batch/swarming use case, we will lose
metrics for jobs that finish before this interval. To handle that,
monitor.stop will be called before the bot exits.

Finally, sigterm is handled in run_bot to avoid losing metrics when
instances get preempted

This PR is part of #4271.