Allowing id token on oauth for appengine (#4314)
### Motivation

In order to get healthchecks on appengine handlers, we need to use GCP
Uptimes. They authenticate through an [oauth id
token](https://cloud.google.com/monitoring/uptime-checks#create). As
things currently stand, oauth on appengine only supportes access tokens,
so this PR solves that.


### Alternatives considered

It would be ideal if we only did one api call, either to validade an id
or access token. However, the
[specification](https://auth0.com/docs/secure/tokens/id-tokens/id-token-structure)
for oauth2 does not present a standard way of, given a token,
differentiating between the two. For that reason, two api calls to GCP
are made.

Part of #4271