Harmonize trusted host and normal corpus pruning code. (#4308)
Get rid of CorpusCrash. Use CrashInfo everywhere.
Using CrashInfo in some places breaks oss-fuzz corpus pruning when there
is a crash in the corpus, because we expect to be using the unit_name
which CrashInfo doesn't have.

It was kind of crazy to have two totally different objects passed to the
same function (only similarity is basically ducktyped).