Instrumenting build retrieval and unpacking times (#4339)
### Motivation

We currently lack metrics for build retrieval and unpacking times. This
PR adds that, with granularity by fuzz target and job type.

There are two different implementations for build downloading/unpacking:

- In the Build class, from which RegularBuild, SplitTargetBuild,
FuchsiaBuild and SymbolizedBuild inherit the downloading/unpacking
behavior
- In the CustomBuild class, which implements its own logic

There are two possible cases for downloading/unpacking: clusterfuzz
either downloads the whole build and unpacks it locally, or unpacks it
remotely. This is the case for all build types except CustomBuild, which
does not perform remote unpacking.

For build retrieval over http, we do not track download time. For all
the other cases, it suffices to keep track of start/finish time for
download and unpacking.

Finally, a _build_type is added to the constructor of the Build class,
from which all other inherit. This is used to track the build type
(debug or release), and is only mutated by SymbolizedBuild when
attempting to fetch a debug build.

Part of #4271