Start to schedule fuzz tasks on batch in OSS-Fuzz (#4397)
Start to schedule fuzz tasks on batch in OSS-Fuzz

The scheduler will work differenly in OSS-Fuzz and Chrome. This only
implements and OSS-Fuzz version.
This version will use job and project weights to decide which fuzzing
jobs to schedule. It then adds these tasks
to the queue for other bots to preprocess and then for the
utask_main_scheduler to actually schedule on batch.
For now, we will only do this for 100 CPUs.
1. Add a cron job to run the scheduler every 15 minutes.
2. Improve region handling in batch (still far from complete).
3. Add function for bulk adding of tasks to queue for use by scheduler.
4. Make fuzz tasks less of a priority on batch than others.