[Monitoring] Move histogram metrics to GeometricBucketer (#4432)
### Motivation

Cumulative distribution metrics from the monitoring initiative were
incorrectly set to use the fixed width bucketer, and/or width=0.05 and
max_buckets=20. This caused percentile metrics to cap at 1, which was
wrong behavior.

This PR attempts to fix that by moving them all to Geometric Bucketer,
without the aforementioned limits. It also reverts #4429 , since it
apparently broke triage.py in chrome.

Part of #4271