Handle disfunctional issue handlers properly in triage (#4476)
Fuchsia testcases were referencing monorail, for which the API was
deprecated. This implied on the cronjob failing prematurely, and no
issues being filed on google3.

This PR adds extra logging, so we can troubleshoot this sort of issue
more easily.

Also, under the triage cronjob, the ValueError exception is handled, to
handle gracefully the case where some testcase references an issue
finder that is not present under the issue tracker config.