Update get_task to return the queue of the task
get_task has been updated to return the queue from which the
task was originally obtained. This is requried so that in case
the task cannot be currently executed, it is added back to the
right queue.

Currently observed that a progression task for a old device like
flame from jobs-android queue was being pushed to
jobs-android-mte-pixel8 queue