Use corpus backup as base for corpus. (#4721)
Use corpus backup as base for corpus when fuzzing. This reduces time to
start fuzzing (including signing the URLs in preprocess and downloading
the urls in main) from 5 minutes to 20 seconds in tests I did. It should
make downloads even faster than they were before the batch migration.

TODO: Do this for pruning.