Skip 'rsync_from_disk()' call if the quarantine corpus is empty
We noticed an increase in fuzzers failing in the Coverage Client Job. We observed
that the major cause was a failure during corpus pruning. Calling 'rsync_from_disk()'
with an empty quarantine corpus leads to trigger unintended assertion in the corpus
pruning job, which subsequently causes the Coverage Client Job to fail. To address this,
we now skip the 'rsync_from_disk()' call if the quarantine corpus is empty, preventing
unnecessary failures.