FIX Prevent None values in GCP log labels (#4729)
While locally debugging on my machine for b/403539825, I noticed
`BOT_NAME` was not set. This prevented me from sending logs to GCP, but
the error message wasn't obvious.

To prevent this issue for other developers working with logs in the
future, I suggest we add a default string value (e.g., "null") when
`BOT_NAME` is not available. This should not interfere with production.

---------

Co-authored-by: Carlo Lemos <carlolemos@google.com>