Add script to run grouper locally (#4734)
Adding a script that modifies the grouper code enabling it to be
executed locally.
This is intended to facilitate experimenting and analyzing the grouper
outcome without touching production. There are mainly two features in
it:
* Load and store a snapshot of the testcases attributes available for
grouping, in order to be able to compare different versions of the
grouper against the same state of the testcases.
* Enable executing the grouper with distinct variations, while storing
the results in a graph-like data structure that makes it easier to
inspect groups and undestand why testcases were grouped.

Note: since the code was copied and edited from the original grouper
code, its is better to use the first commit to review changes.