Avoid using the timeout multiplier twice in fuzzing session timeout c… (#4753)
…alculation

This change is a followup to #4746 which was made to ensure the fuzzer's
test timeout was used during a fuzzing session. In the previous change,
if the fuzzer didn't have a timeout set, we still updated the test
timeout (which meant we ended up having the multiplier^2 in the
timeout).