[Fix] Add fuzz target as argument for logs context to avoid access deny to DB (#4757)
After adding structured logs to testcase-based tasks
(https://github.com/google/clusterfuzz/pull/4736), a new error group
started showing up due to insufficient permissions/no current context
for accessing datastore. This is probably due to querying for the
testcase's fuzz target during `utask_main` execution in logs.

This tries to solve it by adding the fuzz target as an argument for the
logs context, since we can retrieve it from the datastore entity in
trusted part and from the protobuf during the untrusted part of the
task.

---------

Signed-off-by: Vinicius da Costa <vtcosta@google.com>