Add CF_TIP_BOOT_FAILED_COUNT metric for cuttlefish boot failures (#4713)
This commit introduces the `CF_TIP_BOOT_FAILED_COUNT` metric to track
failures when booting Cuttlefish with a tip-of-tree build. This metric
is intended to be used as the basis for an alerting mechanism. By
setting up an alerting policy on this metric, we can be notified when
the number of tip-of-tree boot failures exceeds a defined threshold.