Remove unused task-specific logs context (#4769)
During the current implementation of the structured logs, we added
task-specific context managers to try to facilitate adding metadata for
each task. However, this was creating a lot of unused code and
unnecessary complexity, since only the testcase-based context was
actually being used. In addition, this was not actually making it easier
to add specific metadata, since it would still be necessary to handle
the metadata and to add a new log struct.

This PR removes these task-specific context in favor of using the more
generic testcase-based context, and if the need for specific metadata
arises, this should be handled for each task.