[Logging] Add fuzz-related logs context to corpus pruning task (#4774)
### Description
Differently from the tasks instrumented with logs context so far, corpus
pruning (and fuzz task) is not based on a testcase. These two in
particular operate on top of fuzzers, jobs and fuzz targets. This PR:

* Creates a new fuzzing-related logs context based on a given fuzzer,
job type and fuzz target. Since these metadata are the same as in the
testcase-based context (just retrieved from different source), I changed
the testcase context to contain only the testcase ID.
* Added this context manager to corpus pruning task.

### Tests
Since corpus pruning runs on batch, a onebox deployment with candidate
was not viable. Thus, the test was done by running the task's preprocess
locally with the debugger and sending the logs to GCP. Evidence from the
[logs](https://cloudlogging.app.goo.gl/b8o9YHL9V5BoYGV38):

![image](https://github.com/user-attachments/assets/5174cd71-9a27-45b9-a3c5-f1237c9fc94e)