[Logging] Add logs context to Fuzz Task (#4777)
### Description
This PR adds structured logging to the fuzz task by instrumenting it
with the fuzzing-related logs context.
Notes that for preprocess, the fuzz target is set to None until it is is
picked.

### Tests
Onebox deployment to internal (chrome) GCE instance ([link to
logs](https://cloudlogging.app.goo.gl/pWGNvea5jkaDkKm38)):
* Stage - Preprocess:

![image](https://github.com/user-attachments/assets/b4fdfa26-78eb-4997-a566-e8f62da7a0ae)


* Stage - Main:

![image](https://github.com/user-attachments/assets/8b6ddaed-fe49-439c-a353-1e215f2d0602)