[Logging] Add structured log context to Triage (#4787)
### Description
This adds structured logging to the triage job.

To enable this, the triage code was instrumented with the cron log
context and with the testcase-based context, which wraps each triage
execution for a given testcase. To avoid indenting the whole function, a
new private method was created for triaging a single testcase.

Note: grouper will be instrumented in a next PR after
https://github.com/google/clusterfuzz/pull/4781 is merged.

### Test
* Running the cronjob locally with the debugger ([link to
logs](https://cloudlogging.app.goo.gl/dh6JbXqhcSDRkWDK6)).

![image](https://github.com/user-attachments/assets/5afada1c-801e-491f-9111-36962c28936c)