Use correct queue in OSS-Fuzz (#4804)
In OSS-Fuzz, assume the queue for every task to be initially scheduled
in is the preprocess task. Make a config for this so we won't do this
in Chrome or google3.
The incorrect queue was taken from the testcase/platform. Because this
is OSS-Fuzz, these were the project queues.
Nothing reads from these queues anymore because we don't have instance
groups for each project anymore.

Fixes the recent reproducible crashes breakage in oss-fuzz.