[Logging] Instrument Grouper code with structured logs (#4792)
## Description
#### Structured logs for grouper
For the grouper task, there are two distinct cases to be wrapped with
log contexts:
* When a pair of testcases are being compared for grouping: the
grouper-based context is used to wrap each iteration adding both
testcases` ids and group ids into the log entries (this uses the
symmetric logs implementation described below).
* When iterating over single testcases: the testcase-based context was
used, which adds its ID and fuzzer information into log entries.
However, since grouper uses a distinct class for testcases
(`TestcaseAttributes`), the context was adapted to support this as well.

#### Symmetric logs feature
In order to support the case of log entries with paired testcases, a new
feature was implemented enabling generic symmetrical logs, i.e.,
generating multiple log entries from the same emit call, but setting
distinct values for the labels in each entry. This was done by checking
for a specific label in logs extras (`symmetric_logs`), which should
contain a list of extras for each entry.

This should avoid duplicating the log calls and/or having recursive
calls for emit.

## Tests
Changed the grouper code to only retrieve two testcases and emit a log
wrapped with the grouper context . Running this with the debugger
generated the following test evidence:

[Link to logs](https://cloudlogging.app.goo.gl/dZ5prt14iUonmMA7A)

Logs from first testcase:

![image](https://github.com/user-attachments/assets/83145908-b050-4c03-9b59-103d36b08ea7)

Logs from second testcase:

![image](https://github.com/user-attachments/assets/6e2d73a4-8c41-4c99-a962-36108a2ea0d5)