Adding the exporting features for the job-syncer cronjob (#4808)
### Motivation

In order to have testing environments (dev/staging), we need some
mechanism to automatically push workload definitions for the
fuzzers/fuzz targets/etc.

This will be accomplished with the jobs_syncer cronjob, that allows
importing from prod, and exporting to a test enviroment, the following
entities:

* Fuzzer
* Job
* JobTemplate
* DataBundle

These are sufficient, along with their blobs, to kickstart fuzzing in a
Clusterfuzz deployment

### Implementation

This PR will implement the Exporting functionality, in order to dump
entity definitions and their blobs from production to a GCS bucket, from
which these definitions can be exported to the testing environments.

The export will be done to a target bucket, which the cronjob will take
through the EXPORT_BUCKET env var. Within this bucket, the following
layout will be provided:

```
fuzzer/
    some_fuzzer/
        entity.proto
        blobstore_key
        sample_testcase
...
job/
    some_job/
        entity.proto
        custom_binary_key
...
databundle/
    entity.proto
    contents/
        blob1
        blob2
...
jobtemplate/
     entity.proto
```

### Testing strategy

Unit tests to verify serialization, deserialization, blob and entity
copying.

### Refactors

The entity serialization/deserialization logic was moved from the
uworker_io module to the data_types module, so it can be reused without
imposing a dependency from the job exporter onto the uworker module
(unrelated domains)

### Non goals

Importing entities will be adressed in a followup PR