[Tracing] Implement events module (#4807)
Implementing the module responsible for handling clusterfuzz events
(mainly testcase lifecycle events).
This module defines the event data classes and enables emitting these
events throughout the codebase to be persisted in a generic repository.
The default repository is the datastore and testcase lifecycle events
are stored under the `TestcaseLifecycleEvent` entity.

In summary, this PR:
* Creates the generic data model to describe ClusterFuzz events.
* Extends this as base classes for testcase lifecycle events
(implementing one specific event as an example).
* Creates an abstract repository interface for handling and persisting
events.
* Implements the repository methods for Datastore.
* Creates the functions to emit the events, i.e., serialize and store
them into the underlying repository.
* Adds unit tests for this module.