Merge OSS-Fuzz branch into master (#4797)
This merges the OSS-Fuzz branch into master.
The branches diverged when support for fuzzing on batch was added to
OSS-Fuzz.
Among other things this:
Added scheduling of fuzz tasks and made a lot of optimizations to it, in
order to deal with challenges filling up our capacity with fuzzing.
Optimized things like corpus signing, downloading and pruning as
previously these had not been run at OSS-Fuzz scale and levels of
diversity.

---------

Co-authored-by: Vitor Guidi <vitorguidi@gmail.com>
Co-authored-by: Ali HIJAZI <ahijazi@google.com>
Co-authored-by: Vinicius da Costa <vtcosta@google.com>