Schedule fuzz tasks for batch on Chrome. (#4824)
Add support for Chrome which doesn't have project weight and needs a
small quantity of fuzzing jobs on batch for now.
Our config change combined with this will mean we schedule 5 fuzz tasks
on batch every 10 minutes. Tasks have up to 6 hours to complete. This
means a peak utilization of 180 batch tasks/instances or about 360 CPUs.