[Tracing] Emiting `TestcaseRejectionEvent` during analyze task (#4821)
This PR introduces the emits of `TestcaseRejectionEvent` in the desired
steps of the analyze task life cycle as stated in the "Testcase
rejection" section of "Lightweight Clusterfuzz Tracing" doc.

We are also adding tests to the flow of the `handle_noncrash` function
of analyze, including the verification of the emit.

The other cases where `TestcaseRejectionEvent` must be emitted (namely
fuzz task, triage and corpus prunning) will be addressed in separate
PRs.

b/394056013

---------

Co-authored-by: Carlo Lemos <carlolemos@google.com>