[Deployment Revisited][Staging] Add JobTemplate import logic to the job-exporter cronjob (#4831)
### Motivation

In order to kickstart fuzzing in a testing environment, we need to
mirror the Job, Fuzzer and DataBundle entities.
This PR adds the capability of importing Jobs from the exported data,
implemented in https://github.com/google/clusterfuzz/pull/4808 .

### Implementation

Assuming that all data was exported to $SOME_BUCKET, the folder
structure for JobTemplates looks like this:

```
$SOME_BUCKET/
    jobtemplate/
        entities
        some-job-template/
            entity.proto
        another-job-template/
            entity.proto
```

The proto file is the serialized representation of a data_types.Job
entity, and the other files is the associated custom binary, if present.

The entities file contains line separated job template names that were
last exported.

To import said entities, the cronjob will:
* Parse entity names to be imported from the entities file
* Deserialize the protobuf and persist the entity into datastore, by
performing an environment string substitution

Environment string substitution behaves the same way as described in
#4829

Unit tests for:
* Correctly creating a JobTemplate, from the state where it still does
not exist, and performing the desired env string substitution
* Correctly deleting a JobTemplate, when from the state where it is
present in the target environment, in response to the export list not
containing its name
* Correctly updating a JobTemplate, once a newer revision with different
fields is exported from the source project, while also ensuring correct
env string substitution