[Deployment Revisited][Staging] Add DataBundle import logic to the job-exporter cronjob (#4833)
### Motivation

In order to kickstart fuzzing in a testing environment, we need to
mirror the Job, Fuzzer and DataBundle entities.
This PR adds the capability of importing Jobs from the exported data,
implemented in https://github.com/google/clusterfuzz/pull/4808 .

### Implementation

Assuming that all data was exported to $SOME_BUCKET, the folder
structure for data bundles looks like this:

```
$SOME_BUCKET/
    databundle/
        entities
        some-bundle/
            entity.proto
            contents/
              blob1
              blob2
        another-bundle/
            entity.proto
            contents/
              blob3
              blob4
```

The proto file is the serialized representation of a
data_types.DataBundle entity, and the files under the content folder are
the original contents of the data bundle bucket.

The entities file contains line separated data bundle names that were
last exported.

To import said entities, the cronjob will:
* Parse entity names to be imported from the entities file
* RSync the contents folder for the data bundle in question, to a new
bucket in the target project. This is done through the
[get_data_bundle_bucket_name](https://github.com/google/clusterfuzz/blob/ad39a91bd2756a9e84455d6f5de1eb5d9a7ce0ac/src/clusterfuzz/_internal/datastore/data_handler.py#L1026)
function
* Deserialize the protobuf and persist the entity into datastore, while
changing the bucket name field to the rsync target

Unit tests for:
* Correctly creating a DataBundle, from the state where it still does
not exist, and copying over the data bundle contents
* Correctly deleting a DataBundle, when from the state where it is
present in the target environment, in response to the export list not
containing its name
* Correctly updating a DataBundle, once a newer revision with different
fields is exported from the source project, while also ensuring the
blobs remain up to date



### Attention points

The implementation for StorageRSync was incomplete. List_blobs returns
the file names from the root bucket, and we want to use the file names
starting from the contents folder. This worked for exporting the data
bundles, since the blobs were on the top level folder, but not for the
import:

For this reason, the storage rsync implementation was updated. It should
not change the behavior in a real deployment, since it is a unit testing
only class

```
gs://export-bucket/databundle/some-bundle/contents/blob

gsutil ls export-bucket:
databundle/some-bundle/contents/blob

gsutil cp gs://export-bucket/databundle/some-bundle/contents/blob gs://new-bucket/blob
```